✨ feat: add sales return v2 integration

+++ src/main/java/com/br/jfcbxp/rommanel/cdnet/records/requests/CdNetSaleProductRequest.java
@@ -1,6 +1,8 @@
 package com.br.jfcbxp.rommanel.cdnet.records.requests;
 
+import java.math.BigInteger;
+
 public record CdNetSaleProductRequest(String fullProductId,
-                                      Integer quantity
+                                      BigInteger quantity
 ) {
 }
